requirements.txt
requests
influxdb
pillow
paho-mqtt
